# contracts/openapi/diagnosis.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "AgroDiagn\u00f3stico diagnosis",
    "version": "1.0.0",
    "description": "Contrato inicial: salud t\u00e9cnica implementada; operaciones de negocio contract-only. Sin servidores p\u00fablicos inventados."
  },
  "paths": {
...
    "/api/v1/crops": {
      "get": {
        "operationId": "diagnosis_list_crops",
        "summary": "Listado de cultivos activos",
        "description": "Retorna los cultivos activos en el cat\u00e1logo (POTATO, MAIZE). Requiere autenticaci\u00f3n Bearer de usuario.",
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "../schemas/common.schema.json#/$defs/UUID"
            },
            "example": "11111111-1111-4111-8111-111111111111",
            "description": "UUID del cliente o generado por servidor si falta; valor inv\u00e1lido 400; eco en respuesta."
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de cultivos activos.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/CropList"
                },
                "example": {
                  "items": [
                    {
                      "code": "POTATO",
                      "name": "Papa",
                      "active": true,
                      "created_at": "2026-09-30T12:00:00Z",
                      "updated_at": "2026-09-30T12:00:00Z"
                    },
                    {
                      "code": "MAIZE",
                      "name": "Ma\u00edz",
                      "active": true,
                      "created_at": "2026-09-30T12:00:00Z",
                      "updated_at": "2026-09-30T12:00:00Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Autenticaci\u00f3n requerida o token inv\u00e1lido.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "Credenciales ausentes, vencidas o firma inv\u00e1lida.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          }
        },
        "x-status": "implemented",
        "x-owner": "diagnosis",
        "x-increment": 2,
        "x-visibility": "public"
      }
    },
    "/api/v1/crops/{code}/problems": {
      "get": {
        "operationId": "diagnosis_list_crop_problems",
        "summary": "Listado de condiciones candidatas de un cultivo",
        "description": "Retorna las condiciones candidatas activas asociadas a un cultivo activo. Devuelve 404 si el cultivo no existe o est\u00e1 inactivo.",
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "POTATO",
                "MAIZE"
              ]
            },
            "example": "POTATO",
            "description": "C\u00f3digo de cultivo."
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "../schemas/common.schema.json#/$defs/UUID"
            },
            "example": "11111111-1111-4111-8111-111111111111",
            "description": "UUID del cliente o generado por servidor si falta; valor inv\u00e1lido 400; eco en respuesta."
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de problemas del cultivo.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/ProblemList"
                },
                "example": {
                  "items": [
                    {
                      "code": "POTATO_HEALTHY",
                      "crop_code": "POTATO",
                      "name": "Planta sana de papa",
                      "problem_type": "HEALTHY",
                      "model_supported": false,
                      "active": true,
                      "created_at": "2026-09-30T12:00:00Z",
                      "updated_at": "2026-09-30T12:00:00Z"
                    },
                    {
                      "code": "POTATO_EARLY_BLIGHT",
                      "crop_code": "POTATO",
                      "name": "Tiz\u00f3n temprano en papa",
                      "problem_type": "DISEASE",
                      "model_supported": false,
                      "active": true,
                      "created_at": "2026-09-30T12:00:00Z",
                      "updated_at": "2026-09-30T12:00:00Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Autenticaci\u00f3n requerida o token inv\u00e1lido.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "Credenciales ausentes, vencidas o firma inv\u00e1lida.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado o no accesible.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "Cultivo no encontrado o inactivo.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          }
        },
        "x-status": "implemented",
        "x-owner": "diagnosis",
        "x-increment": 2,
        "x-visibility": "public"
      }
    },
    "/api/v1/problems/{code}/recommendations": {
      "get": {
        "operationId": "diagnosis_get_problem_recommendations",
        "summary": "Consultar recomendaciones de una condici\u00f3n agr\u00edcola",
        "description": "Retorna la \u00faltima versi\u00f3n aprobada y activa de la recomendaci\u00f3n para el problema activo especificado. Si no existe contenido aprobado, devuelve 200 con items vac\u00edo. Si el problema o cultivo asociado no existe o est\u00e1 inactivo, devuelve 404.",
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Z][A-Z0-9_]+$",
              "maxLength": 80
            },
            "example": "POTATO_EARLY_BLIGHT",
            "description": "C\u00f3digo de la condici\u00f3n candidata."
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "../schemas/common.schema.json#/$defs/UUID"
            },
            "example": "11111111-1111-4111-8111-111111111111",
            "description": "UUID del cliente o generado por servidor si falta; valor inv\u00e1lido 400; eco en respuesta."
          }
        ],
        "responses": {
          "200": {
            "description": "\u00daltima versi\u00f3n activa aprobada de la recomendaci\u00f3n.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/RecommendationList"
                },
                "example": {
                  "items": [
                    {
                      "id": "22222222-2222-4222-8222-222222222222",
                      "problem_code": "POTATO_EARLY_BLIGHT",
                      "version": 1,
                      "title": "Manejo preventivo de tiz\u00f3n temprano",
                      "summary": "Pr\u00e1cticas culturales y medidas preventivas validadas para el control de tiz\u00f3n temprano.",
                      "cultural_practices": [
                        "Rotaci\u00f3n de cultivos por al menos 2 a\u00f1os.",
                        "Eliminaci\u00f3n de rastrojos infestados tras la cosecha."
                      ],
                      "biological_control": [
                        "Uso de cepas antagonistas de Bacillus subtilis seg\u00fan recomendaci\u00f3n t\u00e9cnica local."
                      ],
                      "preventive_measures": [
                        "Monitoreo semanal del tercio inferior del follaje.",
                        "Evitar riego por aspersi\u00f3n en horas de alta humedad relativa."
                      ],
                      "source_refs": [
                        "Manual de Manejo Integrado de Enfermedades de la Papa, INIA 2024",
                        "Gu\u00eda fitopatol\u00f3gica andina v3"
                      ],
                      "review_reference": "REV-2026-AGRO-001",
                      "reviewed_by": "Ing. Agr\u00f3nomo Certificado",
                      "reviewed_at": "2026-09-30T10:00:00Z",
                      "active": true,
                      "created_at": "2026-09-30T10:00:00Z",
                      "updated_at": "2026-09-30T10:00:00Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Autenticaci\u00f3n requerida o token inv\u00e1lido.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "Credenciales ausentes, vencidas o firma inv\u00e1lida.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado o no accesible.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "Condici\u00f3n agr\u00edcola no encontrada o inactiva.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          }
        },
        "x-status": "implemented",
        "x-owner": "diagnosis",
        "x-increment": 2,
        "x-visibility": "public"
      }
    },
...
    "/api/v1/admin/crops/{code}": {
      "patch": {
        "operationId": "diagnosis_admin_patch_crop",
        "summary": "Actualizar cultivo en el cat\u00e1logo",
        "description": "Solo rol ADMIN. Permite modificar nombre y estado active. El c\u00f3digo es inmutable. Registra auditor\u00eda propia.",
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "POTATO",
                "MAIZE"
              ]
            },
            "example": "POTATO",
            "description": "C\u00f3digo de cultivo inmutable."
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "required": false,
            "schema": {
              "$ref": "../schemas/common.schema.json#/$defs/UUID"
            },
            "example": "11111111-1111-4111-8111-111111111111",
            "description": "UUID del cliente o generado por servidor si falta; valor inv\u00e1lido 400; eco en respuesta."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "../schemas/common.schema.json#/$defs/CropPatch"
              },
              "example": {
                "name": "Papa com\u00fan",
                "active": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cultivo actualizado exitosamente.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Crop"
                },
                "example": {
                  "code": "POTATO",
                  "name": "Papa",
                  "active": true,
                  "created_at": "2026-09-30T12:00:00Z",
                  "updated_at": "2026-09-30T12:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inv\u00e1lida o par\u00e1metros incorrectos.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "examples": {
                  "INVALID_REQUEST": {
                    "value": {
                      "code": "INVALID_REQUEST",
                      "message": "Intento de mutar campos inmutables o cuerpo vac\u00edo.",
                      "correlation_id": "11111111-1111-4111-8111-111111111111"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticaci\u00f3n requerida o token inv\u00e1lido.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "UNAUTHORIZED",
                  "message": "Credenciales ausentes, vencidas o firma inv\u00e1lida.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "403": {
            "description": "Rol insuficiente para ejecutar la operaci\u00f3n.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "Se requiere rol administrativo para esta ruta.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado o no accesible.",
            "headers": {
              "X-Correlation-ID": {
                "required": true,
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/UUID"
                },
                "example": "11111111-1111-4111-8111-111111111111"
              },
              "Cache-Control": {
                "required": true,
                "schema": {
                  "const": "private, no-store"
                },
                "example": "private, no-store"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common.schema.json#/$defs/Error"
                },
                "example": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "Cultivo no encontrado.",
                  "correlation_id": "11111111-1111-4111-8111-111111111111"
                }
              }
            }
          }
        },
        "x-status": "implemented",
        "x-owner": "diagnosis",
        "x-increment": 2,
        "x-visibility": "admin"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "UserBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firma, issuer, audience, expiry, revocaci\u00f3n/estado de cuenta. Solo identidad de usuario."
      },
      "ServiceBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Identidad ai_inference; audience interna dedicada. Bearer de usuario no autoriza. Algoritmo/configuraci\u00f3n en ADR de seguridad antes de incremento 1/3."
      },
      "RefreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-agro_refresh",
        "description": "Cookie HttpOnly; Secure; SameSite=Lax; Path=/api/v1/auth; sin Domain. Rotaci\u00f3n y revocaci\u00f3n."
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Debe coincidir con token ligado a sesi\u00f3n; adem\u00e1s verificar Origin permitido."
      }
    }
  }
}